Settings Login in form

+++ authapp/models.py
@@ -5,7 +5,7 @@ from django.contrib.auth.models import AbstractUser
 
 class User(AbstractUser):
     avatar = models.ImageField(upload_to='users_avatar', blank=True)
-    age = models.PositiveIntegerField(blank=True, null=True)
+    age = models.PositiveSmallIntegerField(blank=True, null=True)
 
     def __str__(self):
         return self.username
